amalgamated processing.utils and utils.utils
