+++ calcification/utils/utils.py
@@ -1,7 +1,10 @@
+import itertools
+import string
 from pathlib import Path
 
 import cbsyst.helpers as cbh
 import numpy as np
+import pandas as pd
 import xarray as xa
 from PIL import Image
 from tqdm.auto import tqdm
@@ -129,3 +132,34 @@ def convert_png_to_jpg(directory: str) -> None:
             # save image as JPG
             rgb_img.save(jpg_file, "JPEG")
             print(f"Converted {png_file} to {jpg_file}")
+
+
+def uniquify_repeated_values(vals: list, uniquify_str: str = "LOC") -> list:
+    """
+    Append a unique suffix to repeated values in a list.
+
+    Parameters:
+        vals (list): List of values.
+
+    Returns:
+        list: List of values with unique suffixes.
+    """
+
+    def zip_letters(letters: list[str]) -> list[str]:
+        """Zip a list of strings with uppercase letters."""
+        al = string.ascii_uppercase
+        return (
+            [f"-{uniquify_str}-".join(i) for i in zip(letters, al)]
+            if len(letters) > 1
+            else letters
+        )
+
+    return [j for _, i in itertools.groupby(vals) for j in zip_letters(list(i))]
+
+
+def safe_to_numeric(col):
+    """Convert column to numeric if possible, otherwise return as is."""
+    try:
+        return pd.to_numeric(col)
+    except (ValueError, TypeError):
+        return col  # return original column if conversion fails
